backend/requirements.txt
django==4.2.5
djangorestframework==3.14.0
pillow==10.0.1
psycopg==3.1.12
psycopg-binary==3.1.12
python-dotenv==1.0.0
